pyproject.toml
[project]
name = "summarizer"
version = "0.1.0"
description = "AI-powered video summarization tool using Gemini and Whisper"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faster-whisper>=1.0.0",
    "torch>=2.0.0",
    "numpy>=1.21.0",
    "google-generativeai>=0.8.0",
    "click>=8.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
